[ir][refactor] Move legacy frontend constructs to frontend.h/cpp
